+++ schema/schema.go
@@ -11,6 +11,16 @@ type AWSCodeCommitConnection struct {
 	SecretAccessKey             string `json:"secretAccessKey"`
 }
 
+// BitbucketServerConnection
+type BitbucketServerConnection struct {
+	Certificate                 string `json:"certificate,omitempty"`
+	GitURLType                  string `json:"gitURLType,omitempty"`
+	InitialRepositoryEnablement bool   `json:"initialRepositoryEnablement,omitempty"`
+	RepositoryPathPattern       string `json:"repositoryPathPattern,omitempty"`
+	Token                       string `json:"token"`
+	Url                         string `json:"url"`
+}
+
 // ContainerOverrides
 type ContainerOverrides struct {
 	Image    string            `json:"image,omitempty"`
@@ -27,6 +37,7 @@ type DatacenterOpsConfiguration struct {
 	DeploymentOverrides     map[string]*DeploymentOverrides `json:"deploymentOverrides,omitempty"`
 	EnableSymbols           bool                            `json:"enableSymbols,omitempty"`
 	ExperimentIndexedSearch bool                            `json:"experimentIndexedSearch,omitempty"`
+	GitoliteIP              string                          `json:"gitoliteIP,omitempty"`
 	GitserverCount          int                             `json:"gitserverCount,omitempty"`
 	GitserverDiskSize       string                          `json:"gitserverDiskSize,omitempty"`
 	GitserverSSH            map[string]string               `json:"gitserverSSH,omitempty"`
@@ -40,14 +51,9 @@ type DatacenterOpsConfiguration struct {
 	LangSwift               bool                            `json:"langSwift,omitempty"`
 	LangTypeScript          bool                            `json:"langTypeScript,omitempty"`
 	NodeSSDPath             string                          `json:"nodeSSDPath,omitempty"`
+	PhabricatorIP           string                          `json:"phabricatorIP,omitempty"`
 	StorageClass            string                          `json:"storageClass,omitempty"`
 	UseRBAC                 bool                            `json:"useRBAC,omitempty"`
-
-	// Dogfood-specific params (should take these out at some point, as they're only for testing purposes)
-
-	PhabricatorIP string `json:"phabricatorIP,omitempty"`
-	GitoliteIP    string `json:"gitoliteIP,omitempty"`
-	PyPIIP        string `json:"pypiIP,omitempty"`
 }
 
 // DeploymentOverrides
@@ -177,74 +183,75 @@ type Settings struct {
 
 // SiteConfiguration Configuration for a Sourcegraph Server site.
 type SiteConfiguration struct {
-	AdminUsernames                 string                     `json:"adminUsernames,omitempty"`
-	AppURL                         string                     `json:"appURL,omitempty"`
-	AuthAllowSignup                bool                       `json:"auth.allowSignup,omitempty"`
-	AuthOpenIDConnect              *OpenIDConnectAuthProvider `json:"auth.openIDConnect,omitempty"`
-	AuthProvider                   string                     `json:"auth.provider,omitempty"`
-	AuthPublic                     bool                       `json:"auth.public,omitempty"`
-	AuthSaml                       *SAMLAuthProvider          `json:"auth.saml,omitempty"`
-	AuthUserIdentityHTTPHeader     string                     `json:"auth.userIdentityHTTPHeader,omitempty"`
-	AuthUserOrgMap                 map[string][]string        `json:"auth.userOrgMap,omitempty"`
-	AutoRepoAdd                    bool                       `json:"autoRepoAdd,omitempty"`
-	AwsCodeCommit                  []AWSCodeCommitConnection  `json:"awsCodeCommit,omitempty"`
-	CorsOrigin                     string                     `json:"corsOrigin,omitempty"`
-	DisableAutoGitUpdates          bool                       `json:"disableAutoGitUpdates,omitempty"`
-	DisableBrowserExtension        bool                       `json:"disableBrowserExtension,omitempty"`
-	DisableBuiltInSearches         bool                       `json:"disableBuiltInSearches,omitempty"`
-	DisableExampleSearches         bool                       `json:"disableExampleSearches,omitempty"`
-	DisablePublicRepoRedirects     bool                       `json:"disablePublicRepoRedirects,omitempty"`
-	DisableTelemetry               bool                       `json:"disableTelemetry,omitempty"`
-	EmailAddress                   string                     `json:"email.address,omitempty"`
-	EmailSmtp                      *SMTPServerConfig          `json:"email.smtp,omitempty"`
-	ExecuteGradleOriginalRootPaths string                     `json:"executeGradleOriginalRootPaths,omitempty"`
-	GitMaxConcurrentClones         int                        `json:"gitMaxConcurrentClones,omitempty"`
-	GitOriginMap                   string                     `json:"gitOriginMap,omitempty"`
-	Github                         []GitHubConnection         `json:"github,omitempty"`
-	GithubClientID                 string                     `json:"githubClientID,omitempty"`
-	GithubClientSecret             string                     `json:"githubClientSecret,omitempty"`
-	GithubEnterpriseAccessToken    string                     `json:"githubEnterpriseAccessToken,omitempty"`
-	GithubEnterpriseCert           string                     `json:"githubEnterpriseCert,omitempty"`
-	GithubEnterpriseURL            string                     `json:"githubEnterpriseURL,omitempty"`
-	GithubPersonalAccessToken      string                     `json:"githubPersonalAccessToken,omitempty"`
-	Gitlab                         []GitLabConnection         `json:"gitlab,omitempty"`
-	GitoliteHosts                  string                     `json:"gitoliteHosts,omitempty"`
-	GitoliteRepoBlacklist          string                     `json:"gitoliteRepoBlacklist,omitempty"`
-	HtmlBodyBottom                 string                     `json:"htmlBodyBottom,omitempty"`
-	HtmlBodyTop                    string                     `json:"htmlBodyTop,omitempty"`
-	HtmlHeadBottom                 string                     `json:"htmlHeadBottom,omitempty"`
-	HtmlHeadTop                    string                     `json:"htmlHeadTop,omitempty"`
-	HttpToHttpsRedirect            bool                       `json:"httpToHttpsRedirect,omitempty"`
-	Langservers                    []Langservers              `json:"langservers,omitempty"`
-	LicenseKey                     string                     `json:"licenseKey,omitempty"`
-	LightstepAccessToken           string                     `json:"lightstepAccessToken,omitempty"`
-	LightstepProject               string                     `json:"lightstepProject,omitempty"`
-	MaxReposToSearch               int                        `json:"maxReposToSearch,omitempty"`
-	NoGoGetDomains                 string                     `json:"noGoGetDomains,omitempty"`
-	OidcClientID                   string                     `json:"oidcClientID,omitempty"`
-	OidcClientSecret               string                     `json:"oidcClientSecret,omitempty"`
-	OidcEmailDomain                string                     `json:"oidcEmailDomain,omitempty"`
-	OidcOverrideToken              string                     `json:"oidcOverrideToken,omitempty"`
-	OidcProvider                   string                     `json:"oidcProvider,omitempty"`
-	Phabricator                    []Phabricator              `json:"phabricator,omitempty"`
-	PhabricatorURL                 string                     `json:"phabricatorURL,omitempty"`
-	PrivateArtifactRepoID          string                     `json:"privateArtifactRepoID,omitempty"`
-	PrivateArtifactRepoPassword    string                     `json:"privateArtifactRepoPassword,omitempty"`
-	PrivateArtifactRepoURL         string                     `json:"privateArtifactRepoURL,omitempty"`
-	PrivateArtifactRepoUsername    string                     `json:"privateArtifactRepoUsername,omitempty"`
-	RepoListUpdateInterval         int                        `json:"repoListUpdateInterval,omitempty"`
-	ReposList                      []Repository               `json:"repos.list,omitempty"`
-	SamlIDProviderMetadataURL      string                     `json:"samlIDProviderMetadataURL,omitempty"`
-	SamlSPCert                     string                     `json:"samlSPCert,omitempty"`
-	SamlSPKey                      string                     `json:"samlSPKey,omitempty"`
-	SearchScopes                   []SearchScope              `json:"searchScopes,omitempty"`
-	SecretKey                      string                     `json:"secretKey,omitempty"`
-	Settings                       *Settings                  `json:"settings,omitempty"`
-	SiteID                         string                     `json:"siteID,omitempty"`
-	SsoUserHeader                  string                     `json:"ssoUserHeader,omitempty"`
-	TlsCert                        string                     `json:"tlsCert,omitempty"`
-	TlsKey                         string                     `json:"tlsKey,omitempty"`
-	TlsLetsencrypt                 string                     `json:"tls.letsencrypt,omitempty"`
-	UpdateChannel                  string                     `json:"update.channel,omitempty"`
-	UseJaeger                      bool                       `json:"useJaeger,omitempty"`
+	AdminUsernames                 string                      `json:"adminUsernames,omitempty"`
+	AppURL                         string                      `json:"appURL,omitempty"`
+	AuthAllowSignup                bool                        `json:"auth.allowSignup,omitempty"`
+	AuthOpenIDConnect              *OpenIDConnectAuthProvider  `json:"auth.openIDConnect,omitempty"`
+	AuthProvider                   string                      `json:"auth.provider,omitempty"`
+	AuthPublic                     bool                        `json:"auth.public,omitempty"`
+	AuthSaml                       *SAMLAuthProvider           `json:"auth.saml,omitempty"`
+	AuthUserIdentityHTTPHeader     string                      `json:"auth.userIdentityHTTPHeader,omitempty"`
+	AuthUserOrgMap                 map[string][]string         `json:"auth.userOrgMap,omitempty"`
+	AutoRepoAdd                    bool                        `json:"autoRepoAdd,omitempty"`
+	AwsCodeCommit                  []AWSCodeCommitConnection   `json:"awsCodeCommit,omitempty"`
+	BitbucketServer                []BitbucketServerConnection `json:"bitbucketServer,omitempty"`
+	CorsOrigin                     string                      `json:"corsOrigin,omitempty"`
+	DisableAutoGitUpdates          bool                        `json:"disableAutoGitUpdates,omitempty"`
+	DisableBrowserExtension        bool                        `json:"disableBrowserExtension,omitempty"`
+	DisableBuiltInSearches         bool                        `json:"disableBuiltInSearches,omitempty"`
+	DisableExampleSearches         bool                        `json:"disableExampleSearches,omitempty"`
+	DisablePublicRepoRedirects     bool                        `json:"disablePublicRepoRedirects,omitempty"`
+	DisableTelemetry               bool                        `json:"disableTelemetry,omitempty"`
+	EmailAddress                   string                      `json:"email.address,omitempty"`
+	EmailSmtp                      *SMTPServerConfig           `json:"email.smtp,omitempty"`
+	ExecuteGradleOriginalRootPaths string                      `json:"executeGradleOriginalRootPaths,omitempty"`
+	GitMaxConcurrentClones         int                         `json:"gitMaxConcurrentClones,omitempty"`
+	GitOriginMap                   string                      `json:"gitOriginMap,omitempty"`
+	Github                         []GitHubConnection          `json:"github,omitempty"`
+	GithubClientID                 string                      `json:"githubClientID,omitempty"`
+	GithubClientSecret             string                      `json:"githubClientSecret,omitempty"`
+	GithubEnterpriseAccessToken    string                      `json:"githubEnterpriseAccessToken,omitempty"`
+	GithubEnterpriseCert           string                      `json:"githubEnterpriseCert,omitempty"`
+	GithubEnterpriseURL            string                      `json:"githubEnterpriseURL,omitempty"`
+	GithubPersonalAccessToken      string                      `json:"githubPersonalAccessToken,omitempty"`
+	Gitlab                         []GitLabConnection          `json:"gitlab,omitempty"`
+	GitoliteHosts                  string                      `json:"gitoliteHosts,omitempty"`
+	GitoliteRepoBlacklist          string                      `json:"gitoliteRepoBlacklist,omitempty"`
+	HtmlBodyBottom                 string                      `json:"htmlBodyBottom,omitempty"`
+	HtmlBodyTop                    string                      `json:"htmlBodyTop,omitempty"`
+	HtmlHeadBottom                 string                      `json:"htmlHeadBottom,omitempty"`
+	HtmlHeadTop                    string                      `json:"htmlHeadTop,omitempty"`
+	HttpToHttpsRedirect            bool                        `json:"httpToHttpsRedirect,omitempty"`
+	Langservers                    []Langservers               `json:"langservers,omitempty"`
+	LicenseKey                     string                      `json:"licenseKey,omitempty"`
+	LightstepAccessToken           string                      `json:"lightstepAccessToken,omitempty"`
+	LightstepProject               string                      `json:"lightstepProject,omitempty"`
+	MaxReposToSearch               int                         `json:"maxReposToSearch,omitempty"`
+	NoGoGetDomains                 string                      `json:"noGoGetDomains,omitempty"`
+	OidcClientID                   string                      `json:"oidcClientID,omitempty"`
+	OidcClientSecret               string                      `json:"oidcClientSecret,omitempty"`
+	OidcEmailDomain                string                      `json:"oidcEmailDomain,omitempty"`
+	OidcOverrideToken              string                      `json:"oidcOverrideToken,omitempty"`
+	OidcProvider                   string                      `json:"oidcProvider,omitempty"`
+	Phabricator                    []Phabricator               `json:"phabricator,omitempty"`
+	PhabricatorURL                 string                      `json:"phabricatorURL,omitempty"`
+	PrivateArtifactRepoID          string                      `json:"privateArtifactRepoID,omitempty"`
+	PrivateArtifactRepoPassword    string                      `json:"privateArtifactRepoPassword,omitempty"`
+	PrivateArtifactRepoURL         string                      `json:"privateArtifactRepoURL,omitempty"`
+	PrivateArtifactRepoUsername    string                      `json:"privateArtifactRepoUsername,omitempty"`
+	RepoListUpdateInterval         int                         `json:"repoListUpdateInterval,omitempty"`
+	ReposList                      []Repository                `json:"repos.list,omitempty"`
+	SamlIDProviderMetadataURL      string                      `json:"samlIDProviderMetadataURL,omitempty"`
+	SamlSPCert                     string                      `json:"samlSPCert,omitempty"`
+	SamlSPKey                      string                      `json:"samlSPKey,omitempty"`
+	SearchScopes                   []SearchScope               `json:"searchScopes,omitempty"`
+	SecretKey                      string                      `json:"secretKey,omitempty"`
+	Settings                       *Settings                   `json:"settings,omitempty"`
+	SiteID                         string                      `json:"siteID,omitempty"`
+	SsoUserHeader                  string                      `json:"ssoUserHeader,omitempty"`
+	TlsCert                        string                      `json:"tlsCert,omitempty"`
+	TlsKey                         string                      `json:"tlsKey,omitempty"`
+	TlsLetsencrypt                 string                      `json:"tls.letsencrypt,omitempty"`
+	UpdateChannel                  string                      `json:"update.channel,omitempty"`
+	UseJaeger                      bool                        `json:"useJaeger,omitempty"`
 }